+++ src/main/java/ru/job4j/ood/core/coresrp/Toy.java
@@ -1,4 +1,4 @@
-package ru.job4j.ood.srp.core;
+package ru.job4j.ood.core.coresrp;
 
 public class Toy {
